# k_thread_abort() cleanups and ARM latency fix