--- a/arch/arm/core/aarch32/cortex_m/thread_abort.c
+++ b/arch/arm/core/aarch32/cortex_m/thread_abort.c
@@ -28,11 +28,7 @@ extern void z_thread_single_abort(struct k_thread *thread);
 
 void z_impl_k_thread_abort(k_tid_t thread)
 {
-	__ASSERT(!(thread->base.user_options & K_ESSENTIAL),
-		 "essential thread aborted");
-
 	z_thread_single_abort(thread);
-	z_thread_monitor_exit(thread);
 
 	if (_current == thread) {
 		if (arch_is_in_isr()) {