- Move some copypasta in `k_thread_abort()` implementations to `z_thread_single_abort()`, which now has a prototype in ksched.h
- Document and clarify what's going on in the ARM implementation and why it is different
- Don't lock interrupts in the ARM implementation, improving latency, since the vanilla implementation doesn't do it