@nashif 2020-09-02
nice, quick benchmark:
before:

`Time to abort a thread (self abort)                         :     680 cycles ,     5666 ns`

after:

`Time to abort a thread (self abort)                         :     650 cycles ,     5416 ns`