@ioannisg 2020-09-02
> nice, quick benchmark:
> before:
> 
> `Time to abort a thread (self abort) : 680 cycles , 5666 ns`
> 
> after:
> 
> `Time to abort a thread (self abort) : 650 cycles , 5416 ns`

@nashif this could be explained easily in ARM, since the lock/unlock routines (which are now removed) introduce synchronization barriers.